Turn rpmts->nrefs into atomic_int
Throughout the code base reference counting is done in a thread-unsave
way.

Turn nrefs into atomic_int so it is thread-save.

Make sure to only use ++ and -- operators and avoid non atomic
combinations of reading and writing.  As the nref can't change again
after reaching zero this is thread-save even as we look at the result
an instruction later. Drop all the static *Unlink functions.

Technically all the *Free() functions should *Link() the instance after the
ref count has reached 0 and the decision to free it has been made.
Otherwise the cleanup code linking and freeing to will lead to a
recusions. Right now only rpmts needs that. So we only do that here.

This does not make the data structures thread save on its own. But it
gives a good foundation.